+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.18.0](https://github.com/CrowdHailer/raxx/tree/0.18.0) - 2018-10-30
+
+### Changed
+
+- `Raxx.set_header/2` now allows duplicate headers
+- `Raxx.get_header/3` converts duplicate header values into a comma-separated list a la RFC7230 3.2.2 
+
 ## [0.17.0](https://github.com/CrowdHailer/raxx/tree/0.17.0) - 2018-10-28
 
 ### Added